Move test only classes from sharedtest package
