+++ app/src/test/java/com/google/android/ground/TestAuthenticationModule.kt
@@ -17,7 +17,7 @@ package com.google.android.ground
 
 import com.google.android.ground.system.auth.AuthenticationManager
 import com.google.android.ground.system.auth.AuthenticationModule
-import com.sharedtest.system.auth.FakeAuthenticationManager
+import com.google.android.ground.system.auth.FakeAuthenticationManager
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.components.SingletonComponent
