+++ app/src/test/java/com/google/android/ground/domain/usecase/SyncSurveyUseCaseTest.kt
@@ -18,12 +18,12 @@ package com.google.android.ground.domain.usecase
 
 import app.cash.turbine.test
 import com.google.android.ground.BaseHiltTest
+import com.google.android.ground.FakeData.SURVEY
 import com.google.android.ground.domain.usecases.survey.SyncSurveyUseCase
 import com.google.android.ground.persistence.local.stores.LocalSurveyStore
+import com.google.android.ground.persistence.remote.FakeRemoteDataStore
 import com.google.android.ground.repository.LocationOfInterestRepository
 import com.google.common.truth.Truth.assertThat
-import com.sharedtest.FakeData.SURVEY
-import com.sharedtest.persistence.remote.FakeRemoteDataStore
 import dagger.hilt.android.testing.BindValue
 import dagger.hilt.android.testing.HiltAndroidTest
 import javax.inject.Inject
