+++ app/src/test/java/com/google/android/ground/persistence/remote/FakeRemoteDataStore.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.sharedtest.persistence.remote
+package com.google.android.ground.persistence.remote
 
 import com.google.android.ground.model.Survey
 import com.google.android.ground.model.SurveyListItem
@@ -22,7 +22,6 @@ import com.google.android.ground.model.User
 import com.google.android.ground.model.locationofinterest.LocationOfInterest
 import com.google.android.ground.model.mutation.Mutation
 import com.google.android.ground.model.toListItem
-import com.google.android.ground.persistence.remote.RemoteDataStore
 import javax.inject.Inject
 import javax.inject.Singleton
 import kotlinx.coroutines.flow.Flow
