+++ app/src/test/java/com/google/android/ground/persistence/remote/firebase/protobuf/ProtobufToFirestoreExtTest.kt
@@ -16,6 +16,7 @@
 
 package com.google.android.ground.persistence.remote.firebase.protobuf
 
+import com.google.android.ground.TimberTestRule
 import com.google.android.ground.proto.Survey
 import com.google.android.ground.proto.Task
 import com.google.android.ground.proto.Task.DateTimeQuestion.Type.BOTH_DATE_AND_TIME
@@ -28,7 +29,6 @@ import com.google.android.ground.test.deeplyNestedTestObject
 import com.google.android.ground.test.nestedTestObject
 import com.google.android.ground.test.testDocument
 import com.google.common.truth.Truth
-import com.sharedtest.TimberTestRule
 import org.junit.ClassRule
 import org.junit.Test
 import org.junit.runner.RunWith
