+++ app/src/test/java/com/google/android/ground/persistence/sync/MediaUploadWorkerTest.kt
@@ -23,8 +23,16 @@ import androidx.work.WorkerFactory
 import androidx.work.WorkerParameters
 import androidx.work.testing.TestListenableWorkerBuilder
 import com.google.android.ground.BaseHiltTest
+import com.google.android.ground.FakeData
 import com.google.android.ground.model.mutation.Mutation
-import com.google.android.ground.model.mutation.Mutation.SyncStatus.*
+import com.google.android.ground.model.mutation.Mutation.SyncStatus.COMPLETED
+import com.google.android.ground.model.mutation.Mutation.SyncStatus.FAILED
+import com.google.android.ground.model.mutation.Mutation.SyncStatus.IN_PROGRESS
+import com.google.android.ground.model.mutation.Mutation.SyncStatus.MEDIA_UPLOAD_AWAITING_RETRY
+import com.google.android.ground.model.mutation.Mutation.SyncStatus.MEDIA_UPLOAD_IN_PROGRESS
+import com.google.android.ground.model.mutation.Mutation.SyncStatus.MEDIA_UPLOAD_PENDING
+import com.google.android.ground.model.mutation.Mutation.SyncStatus.PENDING
+import com.google.android.ground.model.mutation.Mutation.SyncStatus.UNKNOWN
 import com.google.android.ground.model.mutation.SubmissionMutation
 import com.google.android.ground.model.submission.PhotoTaskData
 import com.google.android.ground.model.submission.ValueDelta
@@ -34,13 +42,12 @@ import com.google.android.ground.persistence.local.stores.LocalLocationOfInteres
 import com.google.android.ground.persistence.local.stores.LocalSubmissionStore
 import com.google.android.ground.persistence.local.stores.LocalSurveyStore
 import com.google.android.ground.persistence.local.stores.LocalUserStore
+import com.google.android.ground.persistence.remote.FakeRemoteDataStore
+import com.google.android.ground.persistence.remote.FakeRemoteStorageManager
 import com.google.android.ground.repository.MutationRepository
 import com.google.android.ground.repository.UserMediaRepository
 import com.google.common.truth.Truth.assertWithMessage
 import com.google.firebase.crashlytics.FirebaseCrashlytics
-import com.sharedtest.FakeData
-import com.sharedtest.persistence.remote.FakeRemoteDataStore
-import com.sharedtest.persistence.remote.FakeRemoteStorageManager
 import dagger.hilt.android.testing.BindValue
 import dagger.hilt.android.testing.HiltAndroidTest
 import javax.inject.Inject
