+++ app/src/test/java/com/google/android/ground/repository/TermsOfServiceRepositoryTest.kt
@@ -16,12 +16,12 @@
 package com.google.android.ground.repository
 
 import com.google.android.ground.BaseHiltTest
+import com.google.android.ground.FakeData
 import com.google.android.ground.persistence.remote.DataStoreException
+import com.google.android.ground.persistence.remote.FakeRemoteDataStore
 import com.google.android.ground.system.NetworkManager
 import com.google.common.truth.Truth.assertThat
 import com.google.firebase.firestore.FirebaseFirestoreException
-import com.sharedtest.FakeData
-import com.sharedtest.persistence.remote.FakeRemoteDataStore
 import dagger.hilt.android.testing.BindValue
 import dagger.hilt.android.testing.HiltAndroidTest
 import javax.inject.Inject
