+++ settings.gradle
@@ -18,4 +18,4 @@ plugins {
     id 'org.gradle.toolchains.foojay-resolver-convention' version '0.4.0'
 }
 
-include ':app', ':sharedTest', ':e2eTest'
+include ':app', ':e2eTest'
