--- sharedTest/build.gradle
@@ -1,75 +1,0 @@
-/*
- * Copyright 2022 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-
-apply from: '../config/checkstyle/checkstyle.gradle'
-apply from: '../config/lint/lint.gradle'
-apply from: '../config/jacoco/jacoco.gradle'
-
-android {
-    namespace 'com.sharedtest'
-    compileSdk rootProject.androidCompileSdk
-
-    defaultConfig {
-        minSdkVersion rootProject.androidMinSdk
-        targetSdkVersion rootProject.androidTargetSdk
-    }
-
-    buildTypes {
-        staging {
-        }
-    }
-    flavorDimensions "backend"
-    productFlavors {
-        local {
-            dimension "backend"
-        }
-        dev {
-            dimension "backend"
-        }
-        sig {
-            dimension "backend"
-        }
-        ecam {
-            dimension "backend"
-        }
-    }
-}
-
-dependencies {
-    implementation project(':app')
-    // Room
-    implementation "androidx.room:room-runtime:$roomVersion"
-
-    // Guava
-    implementation 'com.google.guava:guava:33.0.0-jre'
-
-    // Hilt
-    implementation 'androidx.hilt:hilt-work:1.2.0'
-
-    // Firebase
-    implementation platform('com.google.firebase:firebase-bom:33.1.2')
-    implementation 'com.google.firebase:firebase-firestore'
-
-    implementation "junit:junit:$junitVersion"
-    implementation "com.jakewharton.timber:timber:$timberVersion"
-}
-
-kotlin {
-    jvmToolchain rootProject.jvmToolchainVersion
-}
